Ensure consistent URLs if the content is the same
Currently, all URLs use a random UUID, so that a URL for a host changes
when the controller is restarted. It only works well, because the BMH
controller does not track URL changes.

To enable the BMH controller to make decisions based on URLs, this
change switches to stable URLs formed as:

    /<sha256 of the base image>/<sha256 of ignition>

This way, the URL only changes if the base image is updated (e.g. on OCP
upgrade) or new Ignition is used.

In the static server case, the URL is the same as before.